src/backend/chunker/requirements.txt
nats-py==2.7.2
protobuf==4.25.3
python-dotenv==1.0.1
webdriver-manager==4.0.1
selenium==4.21.0
beautifulsoup4==4.12.3